SignSell & SignBuy and buxfix
Modified SignShop in SignBuy
Added SignSell with other features:

Added the possibility of creating signs with infinite space (root
permissions).
Added the ability to create cartels without money (root permissions).
Added command /sign empty to empty the SignSell.

Added support for 63 different languages! To change the language execute
the command /sign lang <code_of_lang> (the code of your language can be
found here)
(This plugin calls the website of the developer to download translations
of languages)

Added message when you try to run the command /sign set <...> on a sign
not registered.
Added event that regenerate signs when loading a new world.

Bugfix:

direction WallSign;
isChunkGenerated();
isOnlinePlayer();
save();
isEnabled();
update files from the old version to the new;
many bug with permission;

Added check so as not to exceed the maximum stack of the item entered in
the sign;
Added protection from the plugin SignModifer;